Fixed a bug where the wrap guide position was incorrect.
This commit updates the code which draws our line wrap guide to position
it correctly. This is done using some new font metrics functions.

Additionally, this adds some extra debugging code and unit tests which
help ensure that our wrap guide offset computations are done correctly.

+++ src/editor/QomposeDecoratedTextEdit.h
@@ -102,7 +102,7 @@ class QomposeDecoratedTextEdit : public QPlainTextEdit
 
 		void gutterPaintEvent(QPaintEvent *e);
 		int gutterWidth();
-		int wrapGuideOffset();
+		qreal wrapGuideOffset();
 
 	public Q_SLOTS:
 		void fullUpdate();
